hm-order-ai-assistant/requirements.txt
openai-agents
openai
python-dotenv
sqlalchemy
aiosqlite
nest_asyncio
streamlit